feat: Implement deep parsing of entity data
This commit introduces the capability to perform a deep parse of the entity data within the save file body. This is controlled by a new `deepParse` flag in the `parseBodyOutline` function. When enabled, the parser will now deserialize the full entity data for each object, in addition to the structural outline. This lays the groundwork for displaying and editing individual game objects.

This also includes the BDD test that verifies this new deep-parsing functionality.